requirements.txt
bcrypt>=4.0.0
argon2-cffi>=21.3.0
